requirements.txt
asyncio
tinkoff-investments
flask
openai